requirements.txt
torch
scikit-learn
pandas
numpy
torch-geometric
pyyaml
libauc==1.3.0
